fsdp(perf): hand bf16 logits to the cross-entropy, drop the full-vocab fp32 buffer
The FSDP actor did logits = model(...).logits.float(), materializing the entire full-vocab
logits tensor [1, T, V] in fp32 (~5GB at T=16384, V=151936) just to reduce it to per-token
logprobs. Keep the model-native bf16 logits and upcast only the per-response chunk to fp32
inside calculate_log_probs_and_entropy (the fused cross-entropy + entropy still run in fp32).

Bit-identical: bf16->fp32 is a lossless upcast, so the fused-CE output and gradients are unchanged
(lp_diff goldens hold). .to(float32, copy=True) replaces .clone() -- a no-op-dtype copy for the
megatron path (which already passes fp32, so that backend is untouched) and an upcast-copy for the
FSDP bf16 path. Frees the full-vocab fp32 logit buffer; only small per-chunk fp32 copies remain.

+++ miles/backends/experimental/fsdp_utils/actor.py
@@ -396,7 +396,10 @@ class FSDPTrainRayActor(TrainRayActor):
                         )
 
                         model_args = self._get_model_inputs_args(batch)
-                        logits = active_model(**model_args).logits.float()
+                        # Keep logits in the model's native dtype (bf16); the per-response chunk is
+                        # upcast to fp32 inside calculate_log_probs_and_entropy. Avoids materializing
+                        # the full-vocab fp32 logits tensor ([1, T, V] ~ 5GB) just to reduce to logprobs.
+                        logits = active_model(**model_args).logits
 
                         result = get_log_probs_and_entropy(
                             logits=logits,
@@ -567,7 +570,8 @@ class FSDPTrainRayActor(TrainRayActor):
     def _train_step(self, batch, step_id, num_microbatches):
         # Prepare model inputs
         model_args = self._get_model_inputs_args(batch)
-        logits = self.model(**model_args).logits.float()
+        # bf16 logits (see log_probs phase); per-response chunks are upcast to fp32 in the loss path.
+        logits = self.model(**model_args).logits
 
         loss, normalizer, log_dict = loss_function(
             args=self.args,
